+++ src/admin_ui/src/admin/widgets/index.ts
@@ -2,3 +2,4 @@ export * from "./json_field";
 export * from "./json_input";
 export * from "./mapping_field";
 export * from "./mapping_input";
+export * from "./simple_array_field";