+++ src/admin_ui/src/admin/widgets/simple_array_field.tsx
@@ -1,0 +1,34 @@
+import {useRecordContext} from "react-admin";
+import Chip from '@mui/material/Chip'
+import {ReactElement} from "react";
+
+
+export interface SimpleArrayFieldProps {
+    source: string;
+    break_lines?: boolean;
+}
+
+export const SimpleArrayField = (props: SimpleArrayFieldProps) => {
+    const {
+        source,
+        break_lines = false,
+    } = props;
+    const record = useRecordContext();
+    const array = record[source]
+    if (typeof array === 'undefined' || array === null || array.length === 0) {
+        return <div/>
+    } else {
+        let elements: ReactElement[] = [];
+        array.forEach((item: any, index: number) => {
+                elements.push(<Chip label={item} key={item}/>);
+                if (break_lines && index < array.length - 1) {
+                    elements.push(<br/>)
+                }
+            }
+        );
+        return <>{elements}</>
+    }
+}
+
+export default SimpleArrayField;
+